blog delete and some starter styles
