+++ apps/client/app/components/Elements/Button/button.css
@@ -10,7 +10,6 @@
   font-size: var(--fs-button);
   height: 3rem;
   transition: all var(--transition-duration);
-  margin-block: var(--spacing-300);
 }
 
 .button:hover:enabled,
