+++ apps/client/app/features/blog/components/BlogInfo/blog-info.css
@@ -1,0 +1,4 @@
+.blog-info {
+  display: flex;
+  justify-content: space-between;
+}
