+++ apps/client/app/routes/__auth/__admin/blog/index.tsx
@@ -1,15 +1,22 @@
 import type { DataFunctionArgs, LinksFunction } from '@remix-run/node';
 import { json } from '@remix-run/node';
-import { useFetcher, useLoaderData, useOutletContext } from '@remix-run/react';
+import { useLoaderData, useOutletContext } from '@remix-run/react';
 
 import {
   createSupabaseServerClient,
   getAllBlogs
 } from '~/service/supabase/supabase.service';
-import { Button, links as buttonStyles } from '~/components/Elements/Button';
+import {
+  BlogInfo,
+  links as blogInfoStyles
+} from '~/features/blog/components/BlogInfo';
 import type { AdminContext } from '../../__admin';
+import blogStyles from '~/styles/routes/blog.css';
 
-export const links: LinksFunction = () => [...buttonStyles()];
+export const links: LinksFunction = () => [
+  { rel: 'stylesheet', href: blogStyles },
+  ...blogInfoStyles()
+];
 
 export const loader = async ({ request }: DataFunctionArgs) => {
   const supabase = createSupabaseServerClient(request);
@@ -23,27 +30,15 @@ export const loader = async ({ request }: DataFunctionArgs) => {
 export default function Blog() {
   const { blogs } = useLoaderData<typeof loader>();
   const { isAdmin } = useOutletContext<AdminContext>();
-  const { submit } = useFetcher();
-
-  const handleSubmit = (id: number) => {
-    const formData = new FormData();
-    formData.set('id', id.toString());
-    submit(formData, { method: 'post', action: '/api/blog/delete-blog' });
-  };
 
   return (
     <section>
       <h2>Blog</h2>
-      {blogs?.map(({ title, id }) => (
-        <div key={id}>
-          <div>{title}</div>
-          {isAdmin ? (
-            <Button type="button" width="none" onClick={() => handleSubmit(id)}>
-              &#10005;
-            </Button>
-          ) : null}
-        </div>
-      ))}
+      <div className="blog-container">
+        {blogs?.map(({ title, id }) => (
+          <BlogInfo key={id} isAdmin={isAdmin} id={id} title={title} />
+        ))}
+      </div>
     </section>
   );
 }
