+++ apps/client/styles/routes/blog.css
@@ -1,0 +1,5 @@
+.blog-container {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  grid-gap: var(--spacing-300);
+}
